Project/VirtualCanva/requirements.txt
flask
opencv-python
numpy
mediapipe
cvzone
pillow
google-generativeai
